Revert "Refactor/Learn"
